refactor (heap): rename priority queue to heap
